+++ src/markdown_checker/models/base.py
